[core] Prep work to refactor the GKE code.
This work is preparation for other future work.
